+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 * expose security definition as an enum type
 * expose schema types which define `parameters` with new `ParameterOrRef` type
+* Adds License object
+* Adds Contact object
+* Derives Default for all structs
 
 # 0.1.5